pyproject.toml
[tool.poetry]
name = "discord-bot-google-calendar"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.111.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
pandas = "^2.1.4"
anthropic = "^0.7.7"
openmeteo-requests = "^1.1.0"
requests-cache = "^1.1.1"
retry_requests = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
